assgine ticket automatiquement to a support

+++ src/main/java/com/example/QuickFixersBackend/dto/ticket/TicketResponseDTO.java
@@ -5,6 +5,7 @@ import com.example.QuickFixersBackend.enums.Statut;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -17,7 +18,7 @@ public class TicketResponseDTO {
     private String description;
     private Statut statut;
     private LocalDate dateCreation;
-    private double prix;
+    private BigDecimal prix;
     private Long assignedToId;
     private Long serviceId;
     private Long createdById;
